+++ NIPTool/API/external/api/api_v1/endpoints/index.py
@@ -18,24 +18,3 @@ def index(request: Request):
     return templates.TemplateResponse(
         "index.html", context={"request": request, "current_user": ""}
     )
-
-
-@router.get("/new_user")
-def new_user(request: Request):
-    """Log in view."""
-    return templates.TemplateResponse(
-        "new_user.html", context={"request": request, "current_user": ""}
-    )
-
-
-@router.post("/new_user")
-def new_user(request: Request):
-    """Log in view."""
-    return templates.TemplateResponse(
-        "new_user.html",
-        context={
-            "request": request,
-            "current_user": "",
-            "user_info": request.cookies.get("user_info"),
-        },
-    )
