+++ NIPTool/API/external/api/api_v1/templates/index.html
@@ -29,9 +29,9 @@
                                     <label for="exampleInputPassword1">Password</label>
                                     <input type="password" class="form-control" id="password" name="password">
                                 </div>
-                                <a type="submit" class="btn btn-primary">Submit</a>
-                                <a href="{{ url_for('new_user') }}"  class="btn btn-link">New User</a>
                             </form>
+                            <a type="submit" class="btn btn-primary">Submit</a>
+                            <a href="{{ url_for('new_user') }}"  class="btn btn-link">New User</a>
 
                         </div>
                     </div>
